requirements.txt
infrahouse-core ~=  0.17
pytest-infrahouse ~= 0.11
pytest-timeout ~= 2.1
